requirements.txt
asgiref==3.7.2
Django==4.2.3
django-debug-toolbar==4.1.0
django-filter==23.2
djangorestframework==3.14.0
drf-nested-routers==0.93.4
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
